KVM: arm64: Allow RAS field in ID_AA64PFR0_EL1 writable
When migrating between MtCollins and AmpereOne. The RAS field is differ
but not allowed to be writable.

On AmpereOne, the RAS field is 1. On MtCollins, the RAS field is 0.

Allow RAS field in ID_AA64PFR0_EL1 to be writable should be fine. It
will rely on the userspace to write the correct value into it.

Signed-off-by: Shaoqin Huang <shahuang@redhat.com>