KVM: arm64: Allow BT field in ID_AA64PFR1_EL1 writable
When migrating between MtCollins and AmpereOne. The BT field is diff and
not allowed to be writable, so this can cause the migration failed since
user want to write different BT value into it.

On AmpereOne, the BT field is 1. On MtCollins, the BT field is 0.

Allow BT field in ID_AA64PFR1_EL1 to be writable should be fine. It will
rely on the userspace to write the correct value into it.

Signed-off-by: Shaoqin Huang <shahuang@redhat.com>