KVM: arm64: Allow userspace to change ID_AA64PFR1_EL1
Allow userspace to change the guest-visible value of the register with
different way of handling:

  - No changes to features not virtualized by KVM (MPAM_frac, RAS_frac).

  - No changes to features which is controlled by an internal flag
    (MTE). And the MTE related fields is also unwritable (MTE_frac, MTEX).

  - Allow the fields which have a fixed value to be writable but
    explicitly disable them (SME, RNDR_trap, NMI, GCS, THE, DF2, PFAR).

  - Allow the fields to be writable without any effect (BT, SSBS,
    CSV2_frac).

But there is still one problem, how to handle the CSV2_frac correctly
since it depends on the CSV2 value?
The values 0b0001 and 0b0010 are permitted only when
ID_AA64PFR0_EL1.CSV2 is 0b0001.